+++ aframe-blaze/client/extras/aEvent.js
@@ -1,0 +1,5 @@
+UI.registerHelper('aEvent', function () {
+  setAttributes (this, '_aEvent', EVENT_ATTRIBUTES.concat(COMMON_ATTRIBUTES));
+  return Template._aEvent;
+});
+