+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Meteor A-Frame Package Change Log
 
 ## 0.3.0 - March 23, 2016
 
+* Added `aEvent` helper
 * Added new primitives
   * `aBox`
   * `aCircle`